requirements.txt
fastapi
schedule
python-dotenv
prometheus-clientanyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.0
fastapi==0.79.0
idna==3.3
prometheus-client==0.14.1
pydantic==1.9.1
python-dotenv==0.20.0
requests==2.28.1
schedule==1.1.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.10
